bugfixes and improvements
Charge density calculation has been jitted and parallelised.

As a result, rho is now a grid quantity, and will have the value of the most recent deposit, this could be a single species or all species, so best to do a fresh deposit before any diagnostics to ensure the correct values are in the array.

Added a choice between linear, quadratic and cubic particle shapes for charge density deposition. - Does not affect simulation results, current is still linear deposition only

particles with zero weight are no longer initialised

fixed a bug where particles in the first cell would not deposit the correct current